#1697: extract SYN-cookie reply machinery into cookie_reply.rs
Move SYN_COOKIE_REPLY_PENDING_RESERVE, the SynCookieReply enum,
syn_cookie_reply_budget_available, enqueue_syn_cookie_reply, and the
syn_cookie_reply_tests module out of poll_descriptor/mod.rs into a new
#[inline(never)] cold sibling module poll_descriptor/cookie_reply.rs.

enqueue_syn_cookie_reply runs only on the SYN-cookie challenge path
(ScreenCheckOutcome::SynCookieChallenge), which never fires for
established transit flows, so it is a true cold/exception body. Marking
it #[cold] #[inline(never)] places it in .text.unlikely, away from the
hot ingress loop's cache lines, and forbids the compiler from
interleaving the cookie/DDoS-path body into the per-packet codegen unit.
syn_cookie_reply_budget_available stays #[inline] because it is a tiny
helper called only from the already-cold enqueue body.

The function bodies and tests are byte-for-byte identical to their
previous location; only the enclosing module and the inline attributes
change. The mod.rs imports of WorkerTxPipeline and SynCookieChallenge
are dropped (now only needed inside cookie_reply.rs).

Validation: cargo build clean on the userspace-dp crate.
